finally add manajemen stok supplier

+++ src/com/main/models/supplier/deleteSupplier.java
@@ -7,7 +7,7 @@ import java.sql.SQLException;
 import com.main.models.connectionDatabase;
 
 public class deleteSupplier {
-    public static boolean deleteData(int idSupplier, int quantity) {
+    public static boolean deleteData(int idSupplier, double quantity) {
         boolean data = false;
 
         String query = "UPDATE tbl_data_supplier SET quantity = ?, status = 'Out Of Stock' WHERE idSupplier = ?";
@@ -15,7 +15,7 @@ public class deleteSupplier {
         try (Connection conn = connectionDatabase.getConnection();
                 PreparedStatement state = conn.prepareStatement(query)) {
 
-            state.setInt(1, quantity);
+            state.setDouble(1, quantity);
             state.setInt(2, idSupplier);
 
             if (state.executeUpdate() > 0) {
